refactor: TransactionList + SettingsPage 컴포넌트 분리 (#384, #385)
C1: TransactionList 622→130줄
- SearchMode(260줄) + MonthlyView(200줄) 컴포넌트 분리
- 18개 테스트 추가

C2: SettingsPage 848→220줄
- MyAccountSection(380줄) + AppearanceSection(65줄) + ChangelogSection(60줄) 분리
- 15개 테스트 추가

전체 911 tests passed, 동작 변경 없음
